Add assert_throw and fix assert_nothrow
Previously, assert_nothrow was unusable, because the template paramter
pack would not expand because source_location was defined after it.

Now the user has to explicitly wrap his fn to get source location
details. I don't like this much, but I don't see a better option
right now.

assert_throw will fail the test if no exception of the given type
is thrown. If the AnyException type is supplied, it fails only
if no exception is thrown at all.