Remove unused synchronization and allow constexpr tests
For constexpr function, you can test them with :
TEST_ALL_CONSTEXPR(...) and
TEST_ALL_WITH_FIXTURE_CONSTEXPR(...)

The error reporting is decent, but not perfect. Failed
tests throw an exception, which is invalid in a constexpr
context. So the first failed test will cause compilation to fail.

Because of this, constexpr tests do no test summary, as it's
either fully passed or a test failed and printed the appropriate
call trace.

It's a little hard to see the argument values for failed constexpr
tests. I don't think that's too bad, since they are usually simple.
But a better failure would be nice.