Add verify helper functions.
The function v() and vn() verify equality and inequality
of two arguments. Then, they return the left argument.

This allows one-line computing, checking and assigning of values.
This is especially useful in constexpr tests, since you can immediately
see verification failures and only continue if the verification
succeeds.

There are also _v literal operators and comparison using the && and ^
operators. These are cute, but v() and vn() should be preferred.